% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHRVEasy.R
\name{saveHRVIndices}
\alias{saveHRVIndices}
\title{Save the HRV indices as an excel file}
\usage{
saveHRVIndices(RHRVEasyResultObject, saveHRVIndicesInPath = ".")
}
\arguments{
\item{RHRVEasyResultObject}{An object of class \code{RHRVEasyResult} as returned
by \code{RHRVEasy()}}

\item{saveHRVIndicesInPath}{The path where the HRV indices will be saved as
an excel file. The name of the file is automatically created based on the
groups being compared.}
}
\description{
Save the HRV indices as an excel file
}
